+++ app/views/admin/inbound_api_configurations/index.haml
@@ -31,7 +31,7 @@
           - @confs.each do |conf|
             %tr
               %td= conf.external_system_name
-              %td= conf.internal_system_name
+              %td= conf.internal_system.name
               %td= conf.plain_text_api_key_with_fallback
               %td= conf.version
               %td
