+++ db/warehouse/migrate/20230428145659_add_internal_systems_part2.rb
@@ -1,0 +1,6 @@
+class AddInternalSystemsPart2 < ActiveRecord::Migration[6.1]
+  def up
+    require_relative '../../seed_maker'
+    SeedMaker.new.populate_internal_system_choices
+  end
+end
