+++ drivers/hmis_external_apis/spec/factories/inbound_api_configurations.rb
@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :inbound_api_configuration, class: 'HmisExternalApis::InboundApiConfiguration' do
-    sequence(:internal_system_name) { |n| ['referral', 'involvement', 'somesystem', 'someothersystem'][n % 4] }
+    internal_system
     sequence(:external_system_name) { |n| ['LINK', 'MPER', 'AABB', 'QQRR', 'AAAA'][n % 5] }
   end
 end
