+++ drivers/hmis_external_apis/spec/factories/internal_systems.rb
@@ -1,0 +1,8 @@
+FactoryBot.define do
+  factory :internal_system, class: 'HmisExternalApis::InternalSystem' do
+    sequence(:name) do |n|
+      len = HmisExternalApis::InternalSystem::NAMES.length
+      HmisExternalApis::InternalSystem::NAMES[n % len]
+    end
+  end
+end
